Modify check_unit and pass tests.
