+++ essm/equations/_core.py
@@ -160,9 +160,9 @@ class BaseEquation(Eq):
         **Examples:**
 
         >>> from essm.equations.leaf.energy_water import eq_Rs_enbal, eq_El
-        >>> from essm.equations.leaf.energy_water import eq_Hl, eq_Rll
-        >>> eq_Rs_enbal.subs(eq_El, eq_Hl, eq_Rll)
-        Eq(R_s, E_lmol*M_w*lambda_E + a_sh*epsilon_l*sigm*(T_l**4 - T_w**4) + a_sh*h_c*(-T_a + T_l))
+        >>> from essm.equations.leaf.energy_water import eq_Hl
+        >>> eq_Rs_enbal.subs(eq_El, eq_Hl)
+        Eq(R_s, E_lmol*M_w*lambda_E + R_ll + a_sh*h_c*(-T_a + T_l))
         """
         if len(args) == 1:
             if isinstance(args[0], Eq):
