+++ essm/variables/units.py
@@ -149,67 +149,13 @@ def derive_base_dimension(dim):
     """Derive base dimension of dimension"""
     return functools.reduce(
         operator.mul, (
-            Symbol(d) ** p
+            Dimension(d) ** p
             for d, p in dimsys_SI.get_dimensional_dependencies(dim).items()
         ), 1
     )
 
-def collect_factor_and_basedimension(expr):
-    """Return tuple with factor expression and dimension expression."""
-    from essm.variables._core import BaseVariable, Variable
-    if isinstance(expr, BaseVariable):
-        expr = expr.definition.unit 
-    if isinstance(expr, Quantity):
-        return expr.scale_factor, derive_base_dimension(expr.dimension)
-    elif isinstance(expr, Mul):
-        factor = 1
-        dimension = Dimension(1)
-        for arg in expr.args:
-            arg_factor, arg_dim = collect_factor_and_basedimension(arg)
-            factor *= arg_factor
-            dimension *= arg_dim
-        return factor, dimension
-    elif isinstance(expr, Pow):
-        factor, dim = collect_factor_and_basedimension(expr.base)
-        exp_factor, exp_dim = collect_factor_and_basedimension(expr.exp)
-        if exp_dim.is_dimensionless:
-            exp_dim = 1
-        return factor ** exp_factor, dim ** (exp_factor * exp_dim)
-    elif isinstance(expr, Add):
-        factor, dim = collect_factor_and_basedimension(expr.args[0])           
-        for addend in expr.args[1:]:
-            print(addend)
-            addend_factor, addend_dim = \
-                collect_factor_and_basedimension(addend)
-            print(addend_factor, addend_dim)
-            if dim != addend_dim:
-                print('factor, dim:')
-                print((addend_factor, addend_dim))
-                raise ValueError(
-                    'Dimension of "{0}" is {1}, '
-                    'but it should be {2}'.format(
-                        addend, addend_dim, dim.name))
-            factor += addend_factor
-        return factor, dim
-    elif isinstance(expr, Derivative):
-        factor, dim = collect_factor_and_basedimension(expr.args[0])
-        for independent, count in expr.variable_count:
-            ifactor, idim = collect_factor_and_basedimension(independent)
-            factor /= ifactor**count
-            dim /= idim**count
-        return factor, dim
-    elif isinstance(expr, Function):
-        fds = [collect_factor_and_basedimension(
-            arg) for arg in expr.args]
-        return (expr.func(*(f[0] for f in fds)),
-                expr.func(*(d[1] for d in fds)))               
-    elif isinstance(expr, Dimension):
-        return 1, expr
-    else:
-        return expr, Dimension(1)
-
 __all__ = (
-    'derive_baseunit', 'derive_basedimension', 'derive_unit', 'markdown',
+    'derive_baseunit', 'derive_unit', 'markdown',
     'joule', 'kelvin', 'kilogram', 'meter', 'mole', 'pascal', 'second',
     'unit_symbols', 'watt'
 )
